refactor: 5단계 코멘트 수정
